+++ libs/otel/src/index.ts
@@ -1,0 +1,20 @@
+export {
+  DEFAULT_OTLP_ENDPOINT,
+  parseOtlpHeaders,
+  resolveOtlpConfig,
+  type OtlpConfig,
+} from './config';
+export {
+  DEFAULT_ENVIRONMENT,
+  SERVICE_NAMESPACE,
+  buildResource,
+  type ResourceOptions,
+} from './resource';
+export {
+  DEFAULT_METRIC_EXPORT_INTERVAL_MS,
+  createTelemetry,
+  startTelemetry,
+  type Telemetry,
+  type TelemetryOptions,
+} from './telemetry';
+export { getMeter, getTracer } from './tracer';